[MIG] tracking_manager: Migration to 17.0
