+++ tracking_manager/README.rst
@@ -7,7 +7,7 @@ Tracking Manager
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:a7d5bd552b419ff6e2058562ce30c52387acbe3f01cbd1ec3804a78a9fc5a1c9
+   !! source digest: sha256:f9cc75fa1b0588ea2b6b5dfad5e78e5e8f8c23bd04ae65a1a61ff046d8303410
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
