+++ tracking_manager/__manifest__.py
@@ -6,7 +6,7 @@
     "name": "Tracking Manager",
     "summary": """This module tracks all fields of a model,
                 including one2many and many2many ones.""",
-    "version": "16.0.1.1.0",
+    "version": "17.0.1.0.0",
     "category": "Tools",
     "website": "https://github.com/OCA/server-tools",
     "author": "Akretion, Odoo Community Association (OCA)",
@@ -14,13 +14,7 @@
     "license": "AGPL-3",
     "application": False,
     "installable": True,
-    "depends": [
-        "base",
-        "mail",
-    ],
-    "external_dependencies": {
-        "python": ["odoo_test_helper"],
-    },
+    "depends": ["mail"],
     "data": [
         "views/ir_model_fields.xml",
         "views/ir_model.xml",
