+++ tracking_manager/i18n/it.po
@@ -47,11 +47,6 @@ msgid "Automatic Custom Tracking"
 msgstr "Tracciamento personalizzato automatico"
 
 #. module: tracking_manager
-#: model:ir.model.fields,field_description:tracking_manager.field_ir_model__automatic_custom_tracking_domain
-msgid "Automatic Custom Tracking Domain"
-msgstr "Dominio tracciamento personalizzato automatico"
-
-#. module: tracking_manager
 #: model_terms:ir.ui.view,arch_db:tracking_manager.view_model_form
 msgid "Automatic configuration"
 msgstr "Configurazione automatica"
@@ -83,7 +78,7 @@ msgid "Custom Tracking ON"
 msgstr "Tracciamento personalizzato acceso"
 
 #. module: tracking_manager
-#: model_terms:ir.ui.view,arch_db:tracking_manager.view_model_form
+#: model:ir.model.fields,field_description:tracking_manager.field_ir_model__automatic_custom_tracking_domain
 msgid "Domain"
 msgstr "Dominio"
 
@@ -99,10 +94,10 @@ msgstr "Campi"
 
 #. module: tracking_manager
 #: model:ir.model.fields,help:tracking_manager.field_ir_model__automatic_custom_tracking
-msgid "If tick new field will be automatically tracked if the domain match"
+msgid ""
+"If marked, the fields matching the matched by the domain below will be "
+"automatically tracked for this model."
 msgstr ""
-"Se toccare un nuovo campo viene tracciato automaticamente se il dominio "
-"corrisponde"
 
 #. module: tracking_manager
 #: model:ir.model,name:tracking_manager.model_mail_tracking_value
@@ -148,3 +143,11 @@ msgstr "Aggiorna"
 #: model_terms:ir.ui.view,arch_db:tracking_manager.view_model_form
 msgid "Update fields configuration"
 msgstr "Aggiorna configurazione campi"
+
+#~ msgid "Automatic Custom Tracking Domain"
+#~ msgstr "Dominio tracciamento personalizzato automatico"
+
+#~ msgid "If tick new field will be automatically tracked if the domain match"
+#~ msgstr ""
+#~ "Se toccare un nuovo campo viene tracciato automaticamente se il dominio "
+#~ "corrisponde"
