+++ tracking_manager/i18n/tracking_manager.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
+"Project-Id-Version: Odoo Server 17.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -44,11 +44,6 @@ msgid "Automatic Custom Tracking"
 msgstr ""
 
 #. module: tracking_manager
-#: model:ir.model.fields,field_description:tracking_manager.field_ir_model__automatic_custom_tracking_domain
-msgid "Automatic Custom Tracking Domain"
-msgstr ""
-
-#. module: tracking_manager
 #: model_terms:ir.ui.view,arch_db:tracking_manager.view_model_form
 msgid "Automatic configuration"
 msgstr ""
@@ -80,7 +75,7 @@ msgid "Custom Tracking ON"
 msgstr ""
 
 #. module: tracking_manager
-#: model_terms:ir.ui.view,arch_db:tracking_manager.view_model_form
+#: model:ir.model.fields,field_description:tracking_manager.field_ir_model__automatic_custom_tracking_domain
 msgid "Domain"
 msgstr ""
 
@@ -96,7 +91,9 @@ msgstr ""
 
 #. module: tracking_manager
 #: model:ir.model.fields,help:tracking_manager.field_ir_model__automatic_custom_tracking
-msgid "If tick new field will be automatically tracked if the domain match"
+msgid ""
+"If marked, the fields matching the matched by the domain below will be "
+"automatically tracked for this model."
 msgstr ""
 
 #. module: tracking_manager
