+++ tracking_manager/models/__init__.py
@@ -1,4 +1,5 @@
 from . import mail_thread
 from . import ir_model
+from . import ir_model_fields
 from . import models
 from . import mail_tracking_value
